notes_backend/requirements.txt
asgiref==3.8.1
django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
dotenv==0.9.9
pyjwt==2.9.0
python-dotenv==1.1.0
pytz==2025.2
sqlparse==0.5.3
